feat: convert to EE plugin BREAKING CHANGE: user needs a valid license to use the plugin
