--- src/main/java/io/gravitee/policy/threatprotection/xml/XmlException.java
@@ -1,39 +1,0 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.policy.threatprotection.xml;
-
-/**
- * @author Jeoffrey HAEYAERT (jeoffrey.haeyaert at graviteesource.com)
- * @author GraviteeSource Team
- */
-public class XmlException extends Exception {
-
-    private String key;
-
-    public XmlException(String key, String message) {
-        super(message);
-        this.key = key;
-    }
-
-    public XmlException(String key, String message, Exception e) {
-        super(message, e);
-        this.key = key;
-    }
-
-    public String getKey() {
-        return key;
-    }
-}
