+++ src/main/resources/plugin.properties
@@ -2,7 +2,7 @@ id=xml-threat-protection
 name=XML Threat Protection
 version=${project.version}
 description=Protects against main XML threats
-class=io.gravitee.policy.threatprotection.xml.XmlThreatProtectionPolicy
+class=com.graviteesource.policy.threatprotection.xml.XmlThreatProtectionPolicy
 type=policy
 category=security
 icon=xml-threat-protection.svg
